+++ examples/clientserver3/server.cpp
@@ -230,7 +230,7 @@ public:
     {
         while (str.size())
         {
-            int index = str.find(':');
+            unsigned long index = str.find(':');
             if (index != std::string::npos)
             {
                 return str.substr(0, index);
